requirements.txt
✅ Dependencies

speechrecognition

pyttsx3

pyaudio

requests

pillow

pyjokes